# backend/doc/swagger.json
{
  "swagger": "2.0",
  "info": {},
  "paths": {
...
    "/users/aoutme": {
      "get": {
        "description": "Sign Up",
        "tags": [
          "user"
        ],
        "operationId": "aboutme",
        "responses": {
          "200": {
            "$ref": "#/responses/AboutMeSucces"
          },
          "default": {
            "$ref": "#/responses/GenericResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "JWTClaims": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "integer",
          "format": "uint64"
        },
        "Name": {
          "type": "string"
        },
        "Role": {
          "type": "integer",
          "format": "uint8"
        },
        "Verified": {
          "type": "boolean"
        }
      },
      "x-go-package": "country/domain/entity"
    },
    "Response": {
      "type": "object",
      "properties": {
        "Message": {
          "type": "string"
        }
      },
      "x-go-package": "country/domain/entity"
    },
...
  },
  "responses": {
    "AboutMeSucces": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "User": {
            "$ref": "#/definitions/JWTClaims"
          }
        }
      }
    },
    "GenericResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Response"
      }
    },
...
  }
}